+++ src/disclosure-manager/_DisclosureContainer.jsx
@@ -6,6 +6,7 @@ import { injectIntl } from 'react-intl';
 import ContentContainer from 'terra-content-container';
 
 import DisclosureManagerContext from 'terra-disclosure-manager/lib/DisclosureManagerContext';
+import DisclosureManagerDelegate from 'terra-disclosure-manager/lib/DisclosureManagerDelegate';
 
 import { ApplicationLoadingOverlayProvider } from '../application-loading-overlay';
 import { NavigationPromptCheckpoint, navigationPromptResolutionOptionsShape, getUnsavedChangesPromptOptions } from '../navigation-prompt';
@@ -35,6 +36,18 @@ const DisclosureContainer = injectIntl(({ intl, children, navigationPromptResolu
   const disclosureManager = useContext(DisclosureManagerContext);
   const { NotificationBannerProvider, NotificationBanners } = useNotificationBanners();
   const promptCheckpointRef = useRef();
+  const customRegisterDismissCheckRef = useRef();
+
+  const overrideDisclosureManagerContext = useMemo(() => DisclosureManagerDelegate.clone(disclosureManager, {
+    registerDismissCheck: (check) => {
+      customRegisterDismissCheckRef.current = check;
+
+      /**
+       * Return Promise to align with DisclosureManager's default implementation.
+       */
+      return Promise.resolve();
+    },
+  }), [disclosureManager]);
 
   const defaultPromptOptions = useMemo(() => getUnsavedChangesPromptOptions(intl), [intl]);
 
@@ -48,28 +61,38 @@ const DisclosureContainer = injectIntl(({ intl, children, navigationPromptResolu
   }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
   useEffect(() => {
-    disclosureManager.registerDismissCheck(() => new Promise((resolve, reject) => {
-      if (!promptCheckpointRef.current) {
-        resolve();
-        return;
+    disclosureManager.registerDismissCheck(() => {
+      if (customRegisterDismissCheckRef.current) {
+        return customRegisterDismissCheckRef.current();
       }
 
-      promptCheckpointRef.current.resolvePrompts(navigationPromptResolutionOptions || defaultPromptOptions).then(resolve, reject);
-    }));
+      return new Promise((resolve, reject) => {
+        if (!promptCheckpointRef.current) {
+          resolve();
+          return;
+        }
+
+        promptCheckpointRef.current.resolvePrompts(navigationPromptResolutionOptions || defaultPromptOptions).then(resolve, reject);
+      });
+    });
   }, [defaultPromptOptions, disclosureManager, navigationPromptResolutionOptions]);
 
   return (
-    <ApplicationErrorBoundary>
-      <ApplicationLoadingOverlayProvider>
-        <NavigationPromptCheckpoint ref={promptCheckpointRef}>
-          <ContentContainer header={<NotificationBanners />} fill>
-            <NotificationBannerProvider>
-              {children}
-            </NotificationBannerProvider>
-          </ContentContainer>
-        </NavigationPromptCheckpoint>
-      </ApplicationLoadingOverlayProvider>
-    </ApplicationErrorBoundary>
+    <DisclosureManagerContext.Provider value={overrideDisclosureManagerContext}>
+      <ApplicationErrorBoundary>
+        <ApplicationLoadingOverlayProvider>
+          <NavigationPromptCheckpoint
+            ref={promptCheckpointRef}
+          >
+            <ContentContainer header={<NotificationBanners />} fill>
+              <NotificationBannerProvider>
+                {children}
+              </NotificationBannerProvider>
+            </ContentContainer>
+          </NavigationPromptCheckpoint>
+        </ApplicationLoadingOverlayProvider>
+      </ApplicationErrorBoundary>
+    </DisclosureManagerContext.Provider>
   );
 });