+++ src/terra-dev-site/test/slide-panel-manager/NavigationPrompt.test.jsx
@@ -1,18 +1,11 @@
 import React from 'react';
-import classNames from 'classnames/bind';
 import DisclosureComponent from './DisclosureComponent';
 import SlidePanelManager from '../../../slide-panel-manager';
 
-import styles from './SlidePanelManagerDefault.test.module.scss';
-
-const cx = classNames.bind(styles);
-
-const SlidePanelManagerDefault = () => (
-  <div role="main" className={cx('content-wrapper')}>
-    <SlidePanelManager>
-      <DisclosureComponent identifier="root-component" disclosureType="panel" renderHeaderAdapter />
-    </SlidePanelManager>
-  </div>
+const NavigationPrompt = () => (
+  <SlidePanelManager>
+    <DisclosureComponent identifier="root-component" disclosureType="panel" renderHeaderAdapter />
+  </SlidePanelManager>
 );
 
-export default SlidePanelManagerDefault;
+export default NavigationPrompt;